functions/fn-ingest-jobs/requirements.txt
azure-functions
azure-functions-durable
azure-identity
requests
tenacity
